+++ packages/react-widgets/src/layers/MaskLayer.js
@@ -1,10 +1,10 @@
 import { useSelector } from 'react-redux';
 import { SolidPolygonLayer } from '@deck.gl/layers/typed';
 import { MASK_ID } from '@carto/react-core/';
-import { selectSpatialFilter } from '@carto/react-redux/';
+import { selectValidSpatialFilter } from '@carto/react-redux/';
 
 export default function MaskLayer() {
-  const spatialFilterGeometry = useSelector((state) => selectSpatialFilter(state));
+  const spatialFilterGeometry = useSelector(selectValidSpatialFilter);
   const maskData = !!spatialFilterGeometry
     ? [{ polygon: spatialFilterGeometry?.geometry?.coordinates }]
     : [];